requirements.txt
Django
djangorestframework
Pillow
django-oauth-toolkit

model-bakery